backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.115.13",
    "google-api-python-client>=2.172.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "textblob>=0.19.0",
]
